requirements.txt
pandas
PyQt5
requests
spacy
python-docx
beautifulsoup4
